FINAL CHANGE
I know it’s past the deadline, but I found the problem with telnet. I
needed to put the /r/n sequence back into the host char* before I sent
it to the server. My bad. However, it works fine on firefox without
this.

In addition, I found a workaround using telnet without this that we can
demonstrate in lab:

GET /~harris/test.html HTTP/1.1
Host: www.ics.uci.edu
~~Press RETURN key. Then on new line, type SPACE RETURN

The SPACE, RETURN on the new line successfully sends the /r/n to the
server and the request is successful. If they use telnet in the lab
demonstration, just say they need to press a space before the final
return to get it to work. I doubt we would be docked that many points
for that.